store/scope: Reformat comments to be max 80 chars width.

+++ store/scope/hash.go
@@ -16,16 +16,16 @@ package scope
 
 import "fmt"
 
-// MaxStoreID maximum allowed ID from package store. Doesn't matter whether
-// we have a website, group or store ID. int24 (8388607) size at the moment.
+// MaxStoreID maximum allowed ID from package store. Doesn't matter whether we
+// have a website, group or store ID. int24 (8388607) size at the moment.
 const MaxStoreID int64 = 1<<23 - 1
 
 // DefaultHash default Hash value for Default Scope and ID 0. Avoids typing
 // 		scope.NewHash(DefaultID,0)
 const DefaultHash Hash = Hash(Default)<<24 | 0
 
-// Hash defines a merged Scope with its ID. The ID can either be from
-// a website, group or store.
+// Hash defines a merged Scope with its ID. The ID can either be from a website,
+// group or store.
 type Hash uint32
 
 // If we have need for more store IDs then we can change the underlying types here.
@@ -36,9 +36,8 @@ func (h Hash) String() string {
 	return fmt.Sprintf("Scope(%s) ID(%d)", scp, id)
 }
 
-// Unpack extracts a Scope and its ID from a hash.
-// Returned ID can be -1 when the Hash contains invalid data.
-// An ID of -1 is considered an error.
+// Unpack extracts a Scope and its ID from a hash. Returned ID can be -1 when
+// the Hash contains invalid data. An ID of -1 is considered an error.
 func (h Hash) Unpack() (s Scope, id int64) {
 
 	prospectS := h >> 24
@@ -57,26 +56,26 @@ func (h Hash) Unpack() (s Scope, id int64) {
 	return
 }
 
-// HashMaxSegments maximum supported segments or also known as shards.
-// This constant can be used to create the segmented array in other packages.
-const HashMaxSegments int = 256
+// HashMaxSegments maximum supported segments or also known as shards. This
+// constant can be used to create the segmented array in other packages.
+const HashMaxSegments uint16 = 256
 
 const hashBitAnd Hash = Hash(HashMaxSegments) - 1
 
-// Segment generates an 0 < ID <= 255 from a hash. Only used within an array index
-// to optimize map[] usage in high concurrent situations.
-// Also known as shard. An array of N shards is created, each shard contains its
-// own instance of the cache with a lock. When an item with unique key needs to
-// be cached a shard for it is chosen at first by the function Segment(). After
-// that the cache lock is acquired and a write to the cache takes place.
-// Reads are analogue.
+// Segment generates an 0 < ID <= 255 from a hash. Only used within an array
+// index to optimize map[] usage in high concurrent situations. Also known as
+// shard. An array of N shards is created, each shard contains its own instance
+// of the cache with a lock. When an item with unique key needs to be cached a
+// shard for it is chosen at first by the function Segment(). After that the
+// cache lock is acquired and a write to the cache takes place. Reads are
+// analogue.
 func (h Hash) Segment() uint8 {
 	return uint8(h & hashBitAnd)
 }
 
-// NewHash creates a new merged value. An error is equal to returning 0.
-// An error occurs when id is greater than MaxStoreID or smaller 0.
-// An errors occurs when the Scope is Default and id anything else than 0.
+// NewHash creates a new merged value. An error is equal to returning 0. An
+// error occurs when id is greater than MaxStoreID or smaller 0. An errors
+// occurs when the Scope is Default and id anything else than 0.
 func NewHash(s Scope, id int64) Hash {
 	if id > MaxStoreID || id < 0 {
 		return 0
